Create README.md
This project contains a bunch of algorithms. This project is aimed at developing algorithms to be used at specific instances.
Through the creation of this portfolio, I have gained knowledge in machine learning and I have also learnt the use of tools such as UI/Us, Figma, and also improved my understanding of Microsoft Excel, Word and Access. I have gained skills in the use of Python, Git and Figma. In addition, I have improved my typing and computing  skills and the overall process has made me a highly adaptable, open minded and resilient person.
I was faced with many challenges which included designing the pattern of the Algorithms and time constraint.